move sort by prop to derived page view models
